requirements.txt
fastapi
haystack-ai
sentence-transformers
tiktoken
torch
transformers
uvicorn

#pip install -r requirements.txt
